requirements.txt
playwright>=1.42.0
google-genai>=0.7.0
typer>=0.12.0
python-dotenv>=1.0.0
pyperclip>=1.8.0
pytest>=8.0.0
pytest-mock>=3.12.0
rich>=13.0.0
